certmanager/requirements.txt
psycopg2-binary==2.9.1
requests==2.26.0